Renamed findIndex*
