--- src/findIndexSeries.test.mjs
@@ -1,73 +1,0 @@
-
-import { expect, test } from '@jest/globals'
-import findIndexSeries from './findIndexSeries.mjs'
-import Deferred from './Deferred'
-import { range } from 'itertools'
-
-test('findIndexSeries', async () => {
-  const callCount = {}
-  ;[...range(3)].forEach((i) => { callCount[i] = 0 })
-  const d = new Deferred()
-  const ds = [...range(3)].map(() => new Deferred())
-  const p = findIndexSeries([...range(3)], async (v, i) => {
-    callCount[i] += 1
-    ds[i].resolve()
-    await d.promise
-    return v === 0
-  })
-  await ds[0].promise
-  expect(callCount[0]).toBe(1)
-  expect(callCount[1]).toBe(0)
-  expect(callCount[2]).toBe(0)
-  d.resolve()
-  const res = await p
-  expect(res).toBe(0)
-  expect(callCount[0]).toBe(1)
-  expect(callCount[1]).toBe(0)
-  expect(callCount[2]).toBe(0)
-})
-
-test('findIndexSeries ordered', async () => {
-  const callCount = {}
-  ;[...range(3)].forEach((i) => { callCount[i] = 0 })
-  const d = new Deferred()
-  const ds = [...range(3)].map(() => new Deferred())
-  const p = findIndexSeries([...range(3)], async (v, i) => {
-    callCount[i] += 1
-    ds[i].resolve()
-    if (i === 0) {
-      await d.promise
-    }
-    return true
-  })
-  await ds[0].promise
-  expect(callCount[0]).toBe(1)
-  expect(callCount[1]).toBe(0)
-  expect(callCount[2]).toBe(0)
-  d.resolve()
-  const res = await p
-  expect(res).toBe(0)
-  expect(callCount[0]).toBe(1)
-  expect(callCount[1]).toBe(0)
-  expect(callCount[2]).toBe(0)
-})
-
-test('findIndexSeries error', async () => {
-  const callCount = {}
-  ;[...range(3)].forEach((i) => { callCount[i] = 0 })
-  try {
-    await findIndexSeries([...range(3)], async (v, i) => {
-      callCount[i] += 1
-      if (i === 0) {
-        throw new Error('test')
-      }
-      return true
-    })
-    expect(true).toBe(false)
-  } catch (e) {
-    expect(e.message).toBe('test')
-  }
-  expect(callCount[0]).toBe(1)
-  expect(callCount[1]).toBe(0)
-  expect(callCount[2]).toBe(0)
-})
