+++ src/modern-async.mjs
@@ -12,8 +12,6 @@ export { default as filter } from './filter.mjs'
 export { default as filterGenerator } from './filterGenerator.mjs'
 export { default as find } from './find.mjs'
 export { default as findIndex } from './findIndex.mjs'
-export { default as findIndexLimit } from './findIndexLimit.mjs'
-export { default as findIndexSeries } from './findIndexSeries.mjs'
 export { default as forEach } from './forEach.mjs'
 export { default as forEachLimit } from './forEachLimit.mjs'
 export { default as forEachSeries } from './forEachSeries.mjs'
